Change the shape of the generation region and paper-bag-bug-fix
The old generation region was a cylinder.  It's now a box that is +/- 15 cm
perpendicular to the beam, and through the entire detector along the beam.

This fixed a really bad kinematics and vertex position bug for the
generation.  In the macro file, the beam got the wrong tilt in the XY
plane, and the beam direction was 180 deg wrong (pointing upstream, not
down).  This is traced back to a geometry mistake in the GPS rotations.